reorganizing file directory without unnecessary files
