--- list_function_import.py
@@ -1,3 +1,0 @@
-from sum_list import *
-from min_max_list import *
-from max_difference import *
